ECOMM-42 refactor: name files according to style
