+++ src/components/AuthPage/auth-page.html
@@ -3,7 +3,7 @@
   <div class="auth__content" id="auth-content">
     <div class="auth__form">
       <signup-form class="auth__form" role="form"></signup-form>
-      <signin-form class="auth__form" role="form"></signin-form>
+      <login-form class="auth__form" role="form"></login-form>
     </div>
   </div>
 </div>
