+++ src/components/Header/header.html
@@ -50,9 +50,9 @@
       </div>
       <div class="header__login">
         <div class="login__links">
-            <span data-href="/signin" class="login__link login__link_first">Sign In</span>
+            <span data-href="/login" class="login__link login__link_first">Sign In</span>
             <span class="separator">|</span>
-            <span data-href="/register" class="login__link login__link_second">Join</span>
+            <span data-href="/signup" class="login__link login__link_second">Join</span>
         </div>
           <div data-href="/profile" class="login__button login__button_my-account">
             <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
