+++ src/components/SignupForm/signup-form.html
@@ -107,7 +107,7 @@
     <button id="form-submit" type="submit" class="field__button ">Sign Up</button>
     <p class="field__description">
       Already have an  account?
-      <a href="#" data-href="/login">Log in</a>
+      <a href="#" data-href="#">Log in</a>
     </p>
   </div>
 </form>
