+++ src/styles/styles.ts
@@ -1,4 +1,4 @@
 import bootstrap from './bootstrap.min.module.css';
-import login from '../components/Login/login.module.scss';
+import login from '../components/LoginForm/login-form.module.scss';
 
 export { bootstrap, login };
